+++ next.config.js
@@ -14,12 +14,6 @@ const nextConfig = {
         port: '',
         pathname: '/vi/**',
       },
-      {
-        protocol: 'http',
-        hostname: 'k.kakaocdn.net',
-        port: '',
-        pathname: '/dn/**',
-      },
     ],
   },
