+++ src/app/idea/page.tsx
@@ -1,5 +1,6 @@
 import React from 'react';
 
+
 import { Layout } from '@/components/common';
 import { IdeaList } from '@/components/idea-page';
