+++ src/app/mypage/page.tsx
@@ -1,5 +1,6 @@
 'use client';
 import { Layout } from '@/components/common';
+// import BadgesList from '@/components/mypage/BadgesList';
 import JoinedChallenge from '@/components/mypage/JoinedChallenge';
 import PointTree from '@/components/mypage/PointTree';
 import UserProfile from '@/components/mypage/UserProfile';
@@ -10,11 +11,12 @@ export default function Page({ params: { slug } }: Props) {
   return (
     <Layout>
       {slug}
-      <div className="justify-evenly sm:flex-col md:flex-row lg:flex md:px-2">
-        <div className="md:w-auto sm:w-full flex-shrink-0 md:px-4">
+      <div className="flex justify-evenly">
+        <div>
           <h3 className="mb-4">마이페이지</h3>
           <div className="p-10 bg-white drop-shadow-md">
             <UserProfile />
+            {/* <BadgesList /> */}
             <PointTree />
           </div>
         </div>
