+++ src/app/shared/queries.keys.ts
@@ -4,5 +4,3 @@ export const REVIEWS = 'challenge_certify';
 export const JOIN_CHALLENGE = 'curUser_join_check';
 export const IDEA_COMMENTS = 'idea_comments';
 export const NATURESTORY = 'nature_story';
-export const USER_REVIEWS = 'curUser_challenge_certify';
-export const USER_CHALLENGES = 'curUser_challenges';
