+++ src/components/common/Button.tsx
@@ -23,10 +23,7 @@ export const Button: React.FC<ButtonProps> = ({ children, btnType, size = 'mediu
       buttonSize = 'w-full py-4 text-lg ';
       break;
     case 'large':
-      buttonSize = 'px-[2rem] py-2 text-[1.25rem]';
-      break;
-    case 'large-mo':
-      buttonSize = 'px-[1.875rem] py-2 text-[1rem]';
+      buttonSize = 'px-[2.375rem] py-2 text-[1.25rem]';
       break;
     case 'medium':
       buttonSize = 'px-[1.875rem] py-2 text-base';
