+++ src/components/common/Footer.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 
 export function Footer() {
   return (
-    <footer className="w-full py-8 bg-navy flex justify-center">
+    <footer className="w-full bg-navy py-[39px] px-[562px] flex justify-center">
       <p className="text-sm text-sub6 whitespace-nowrap ">
         Copyright 2023. <strong>PROJECT 13.7</strong> All rights reserved.
       </p>
