+++ src/components/common/Header.tsx
@@ -1,10 +1,9 @@
 'use client';
-import React, { useEffect, useState } from 'react';
+import React, { useEffect } from 'react';
 
 import Image from 'next/image';
 import Link from 'next/link';
 import { usePathname } from 'next/navigation';
-import { AiOutlineAlignRight, AiOutlineClose } from 'react-icons/ai';
 
 import { getUser } from '@/app/api/users';
 import useSessionStore from '@/store/sesson.store';
@@ -19,15 +18,6 @@ import { Layout } from '.';
 export function Header() {
   const { session, isLoaded, setSession } = useSessionStore();
   const params = usePathname();
-  const [isMobileMenuOpen, setIsMobileMenuOpen] = useState(false);
-
-  const toggleMobileMenu = () => {
-    setIsMobileMenuOpen(!isMobileMenuOpen);
-  };
-
-  const closeMobileMenu = () => {
-    setIsMobileMenuOpen(false);
-  };
 
   useEffect(() => {
     const refresh = async () => {
@@ -67,79 +57,39 @@ export function Header() {
     });
   }, []);
   return (
-    <>
-      <div className="w-full sticky top-0 bg-white text-black text-lg z-10">
-        <Layout>
-          <div className="md:flex items-center justify-between hidden px-10 py-8">
-            <div className="flex gap-8">
-              <Link href="/" className="flex font-semibold gap-2">
-                <Image src={logo} alt="logo" />
-                <Image src={logoTitle} alt="logo title" />
-              </Link>
-              <nav className="flex gap-8 ml-8">
-                {navCategory.map(item => (
-                  <Link href={item.pathname} className="text-sub6" key={item.title}>
-                    <h5 className={`${params === item.pathname ? 'text-black' : ''} font-semibold `}>{item.title}</h5>
-                  </Link>
-                ))}
-              </nav>
-            </div>
-            {isLoaded && (
-              <div className="flex gap-4 text-base">
-                {session ? (
-                  <>
-                    <Link href="/mypage" className={`${params === '/mypage' ? 'text-black' : 'text-sub6'} flex gap-2 text-lg font-medium`}>
-                      마이페이지
-                    </Link>
-                    <SignOut />
-                  </>
-                ) : (
-                  <Auth />
-                )}
-              </div>
-            )}
-          </div>
-        </Layout>
-      </div>
-
-      <div className="w-full sticky top-0 px-4 pb-4 bg-white text-sub6 text-lg z-10 font-medium md:hidden shadow-sm justify-center">
-        <div className="flex items-center justify-between">
-          <Link href="/" className="flex font-semibold gap-2" onClick={closeMobileMenu}>
-            <Image src={logo} alt="logo" />
-            <Image src={logoTitle} alt="logo title" />
-          </Link>
-          <button className="font-bold text-xl relative text-right justify-end" onClick={toggleMobileMenu}>
-            {isMobileMenuOpen ? <AiOutlineClose size={30} /> : <AiOutlineAlignRight size={30} />}
-          </button>
-        </div>
-
-        {isMobileMenuOpen && (
-          <div className="md:hidden">
-            <nav className="overflow-y-auto flex flex-col py-10 gap-8 ml-4 justify-center">
+    <div className="w-full sticky top-0 bg-white text-black px-10 py-8 text-lg z-10">
+      <Layout>
+        <div className=" flex items-center justify-between">
+          <div className="flex gap-8">
+            <Link href="/" className="flex font-semibold gap-2">
+              <Image src={logo} alt="logo" />
+              <Image src={logoTitle} alt="logo title" />
+            </Link>
+            <nav className="flex gap-8 ml-8">
               {navCategory.map(item => (
-                <Link href={item.pathname} className="text-sub6 text-left" key={item.title} onClick={closeMobileMenu}>
+                <Link href={item.pathname} className="text-sub6" key={item.title}>
                   <h5 className={`${params === item.pathname ? 'text-black' : ''} font-semibold `}>{item.title}</h5>
                 </Link>
               ))}
             </nav>
-            {isLoaded && (
-              <div className="flex flex-col py-4 gap-4 ml-4 justify-center mb-4" onClick={closeMobileMenu}>
-                {session ? (
-                  <>
-                    <Link href="/mypage" className={`${params === '/mypage' ? 'text-black' : 'text-sub6'} flex gap-2 text-lg font-medium`}>
-                      마이페이지
-                    </Link>
-                    <SignOut />
-                  </>
-                ) : (
-                  <Auth />
-                )}
-              </div>
-            )}
           </div>
-        )}
-      </div>
-    </>
+          {isLoaded && (
+            <div className="flex gap-4 text-base">
+              {session ? (
+                <>
+                  <Link href="/mypage" className={`${params === '/mypage' ? 'text-black' : 'text-sub6'} flex gap-2 text-lg font-medium`}>
+                    마이페이지
+                  </Link>
+                  <SignOut />
+                </>
+              ) : (
+                <Auth />
+              )}
+            </div>
+          )}
+        </div>
+      </Layout>
+    </div>
   );
 }
 
@@ -158,6 +108,6 @@ const navCategory = [
   },
   {
     title: '환경 이야기',
-    pathname: '/nature-story',
+    pathname: '/naturestory',
   },
 ];
