+++ src/components/common/Input.tsx
@@ -11,11 +11,10 @@ interface InputProps extends InputHTMLAttributes<HTMLInputElement> {
 
 export const Input = ({ _size, onChange, type, inputStyle, ...props }: InputProps) => {
   let inputSize: string = '';
+  // 추가필요? background-color style (login modal input bg -> sub3)
   const inputDefault = 'rounded-lg font-normal text-base border border-opacityblack outline-none ';
   const sizeFunc = () => {
     switch (_size) {
-      case 'xs':
-        return (inputSize = 'w-[200px] h-[2rem] py-3 px-2');
       case 'sm':
         return (inputSize = 'w-[388px] h-[3rem] py-3 px-4');
       case 'md':
