+++ src/components/idea-page/IdeaList.tsx
@@ -48,12 +48,10 @@ export function IdeaList() {
                 <IdeaContent key={item.post_id} item={item} />
               ))}
             </div>
-            {hasNextPage && (
-              <Button btnType="borderBlack" size="large" rounded onClick={clickNextPage}>
-                더보기
-                <AiOutlineArrowRight size={20} />
-              </Button>
-            )}
+            <Button btnType="borderBlack" size="large" rounded onClick={clickNextPage}>
+              더보기
+              <AiOutlineArrowRight size={20} />
+            </Button>
           </div>
         </>
       ) : (
