+++ src/components/mypage/UserProfile.tsx
@@ -1,69 +1,200 @@
-import React, { useEffect } from 'react';
+import React from 'react';
 import { useState } from 'react';
 
-import Image from 'next/image';
-import { AiFillEnvironment } from 'react-icons/ai';
+import { useMutation, useQuery } from '@tanstack/react-query';
+import { useDropzone } from 'react-dropzone';
+import { v4 } from 'uuid';
 
-import { getUserProfile } from '@/app/api/mypage';
+import { postUserProfileImg } from '@/app/api/mypage';
+import { Button, Input } from '@/components/common';
 import useSessionStore from '@/store/sesson.store';
 
-import BadgesList from './BadgesList';
-import EditProfile from './EditProfile';
 import RankingGuide from './RankingGuide';
+import { supabase } from '../../../supabase/supabaseConfig';
 
 import type { User } from '@/types/db.type';
-
-const profileDefault = 'https://fvkzqozjdtlaogexuuin.supabase.co/storage/v1/object/public/project/userProfileImg/userProfileDefault?t=2023-09-04T01%3A41%3A06.168Z';
 
 export default function UserProfile() {
   const session = useSessionStore((state: { session: any }) => state.session);
 
   const [editMode, setEditMode] = useState(false);
-  const [userProfile, setUserProfile] = useState<User | null>(null);
+  const [editedProfile, setEditedProfile] = useState<any>(null);
 
-  useEffect(() => {
-    const fetchUserProfile = async () => {
-      const profile = await getUserProfile(session?.user_id);
-      setUserProfile(profile || null);
-    };
+  // useQuery를 사용하여 사용자 프로필 가져오기
+  const { data: userProfile } = useQuery(['userProfile', session?.user_id], async () => {
+    const response = await supabase.from('users').select('*').eq('user_id', session?.user_id);
+    return response.data?.[0];
+  });
 
-    fetchUserProfile();
-  }, [session?.user_id]);
+  // useMutation을 사용하여 사용자 프로필 업데이트
+  const editProfileMutation = useMutation(['editProfile', session?.user_id], async (updatedProfile: User) => {
+    const imgName = v4();
+    try {
+      if (imgFile) {
+        postImgMutation.mutate({ imgFile, imgName });
+      }
+      return supabase.from('users').update(updatedProfile).eq('user_id', session?.user_id);
+    } catch (error) {}
+  });
 
   const handleEditClick = () => {
     setEditMode(true);
+    setEditedProfile((prev: any) => ({
+      ...prev,
+      nickname: userProfile?.nickname || '',
+      address: userProfile?.address || '',
+      profile_img: userProfile?.profile_img || '',
+      user_id: userProfile?.user_id || '',
+      created_at: userProfile?.created_at || '',
+      point: userProfile?.point || 0,
+      email: userProfile?.email || '',
+    }));
   };
 
-  if (!userProfile) {
-    return <div>Loading...</div>;
-  }
+  const handleCancelClick = () => {
+    setEditMode(false);
+    setPreviewImg(undefined);
+  };
+
+  const postImgMutation = useMutation({
+    mutationFn: postUserProfileImg,
+  });
+
+  // 이미지를 저장소에 업로드하고 이미지 URL을 반환하는 함수
+  const uploadImageAndGetUrl = async (imgFile: File, imgName: string) => {
+    await postImgMutation.mutateAsync({ imgFile, imgName });
+
+    // 이미지 URL 받아오기
+    const imgUrlResponse = await supabase.storage.from('project').getPublicUrl(`userProfileImg/${imgName}`);
+    return imgUrlResponse;
+  };
+
+  const handleSaveClick = async () => {
+    if (!editedProfile) {
+      return;
+    }
+    try {
+      if (imgFile && postImgMutation.isSuccess) {
+        const imgName = v4();
+        const imgUrlResponse = await uploadImageAndGetUrl(imgFile, imgName);
+        const imgUrl = imgUrlResponse.data.publicUrl;
+
+        // 유저 프로필 정보 업데이트
+        const updatedProfile = {
+          ...editedProfile,
+          profile_img: imgUrl,
+        };
+
+        await editProfileMutation.mutateAsync(updatedProfile);
+      } else {
+        await editProfileMutation.mutateAsync(editedProfile);
+      }
+
+      setEditMode(false);
+      setImgFile(undefined); // imgFile 상태 초기화
+      setPreviewImg(undefined);
+    } catch (error) {
+      console.error('사용자 프로필 업데이트 오류:', error);
+    }
+  };
+
+  const [imgFile, setImgFile] = useState<File | undefined>(undefined);
+  const [previewImg, setPreviewImg] = useState<string | ArrayBuffer | undefined>(undefined);
+  const { getRootProps, getInputProps } = useDropzone();
+
+  // input에서 사진 첨부 => DB state 할당, 미리보기 state 할당 함수 실행
+  const handleChangeImg = (e: React.ChangeEvent<HTMLInputElement>) => {
+    const selectedFile = e.target.files?.[0];
+    if (selectedFile) {
+      imgUpload(selectedFile);
+    }
+  };
+
+  // Image 취소 버튼 click시 실행
+  const handleCancelImg = () => {
+    setImgFile(undefined);
+    setPreviewImg(undefined);
+  };
+
+  // 첨부된 파일 읽고 DB state 할당, 미리보기 state 할당
+  const imgUpload = (selectedFile: File) => {
+    if (selectedFile) {
+      setImgFile(selectedFile);
+      const reader = new FileReader();
+      reader.readAsDataURL(selectedFile);
+      reader.onloadend = () => {
+        if (reader.result) {
+          setPreviewImg(reader.result);
+        }
+      };
+    }
+  };
 
   return (
     <>
       {editMode ? (
-        <EditProfile setEditMode={setEditMode} userProfile={userProfile} setUserProfile={setUserProfile} />
+        <>
+          <div className="text-center mb-8">
+            <div className="flex flex-col text-center">
+              {typeof previewImg === 'string' ? (
+                <div className="flex flex-col items-center">
+                  <img src={previewImg} alt="Preview Profile Image" width={100} height={100} className="rounded-full inline-block mb-4" />
+                  <button className="text-sm opacity-50 mb-2" onClick={handleCancelImg}>
+                    현재 이미지 삭제
+                  </button>
+                </div>
+              ) : (
+                <>
+                  <div {...getRootProps()}>
+                    <input accept="image/*" type="file" {...getInputProps()} onChange={event => handleChangeImg(event)} />
+                    <img
+                      src={userProfile?.profile_img ? `${userProfile?.profile_img}` : 'https://fvkzqozjdtlaogexuuin.supabase.co/storage/v1/object/public/project/userProfileImg/polar-bear_1f43b-200d-2744-fe0f.png?t=2023-09-04T01%3A38%3A29.808Z'}
+                      alt="profileDefaultImg"
+                      width={100}
+                      height={100}
+                      className="rounded-full inline-block mb-4"
+                    />
+                  </div>
+                </>
+              )}
+              <input type="file" accept="image/*" id="profileImg" className="cursor-pointer" onChange={e => handleChangeImg(e)} />
+            </div>
+            <div className="space-y-1 flex flex-col">
+              <Input type="text" value={editedProfile?.nickname || ''} onChange={e => setEditedProfile((prev: any) => ({ ...prev, nickname: e.target.value }))} placeholder="이름" _size={'sm'} />
+              <Input type="text" value={editedProfile?.address || ''} onChange={e => setEditedProfile((prev: any) => ({ ...prev, address: e.target.value }))} placeholder="주소" _size={'sm'} />
+            </div>
+            <div className="flex gap-2 justify-center mx-auto my-2">
+              <Button btnType={'borderBlack'} size={'small'} onClick={handleCancelClick}>
+                수정 취소
+              </Button>
+              <Button btnType={'black'} size={'small'} onClick={handleSaveClick}>
+                저장하기
+              </Button>
+            </div>
+          </div>
+        </>
       ) : (
         <>
-          <div className="text-center m-4 ">
-            <Image src={userProfile?.profile_img || profileDefault} alt="profileDefaultImg" width={100} height={100} className="w-32 h-32 flex justify-center overflow-hidden object-cover rounded-md mx-auto m-2" />
-          </div>
-          <div className="flex justify-center items-center gap-0.5 p-2">
-            <p className="font-semibold text-lg">{userProfile?.nickname}</p>
-            <RankingGuide />
-          </div>
-          <p className="text-sm opacity-50 flex justify-center gap-1 items-center">
-            <AiFillEnvironment size={15} /> {userProfile?.address}
-          </p>
-          <div className="flex justify-center items-center my-3">
-            <button className="border-sub5 text-sub6 px-4 py-1 gap-2 border rounded-md text-sm mb-2 flex justify-center items-center" onClick={handleEditClick}>
-              프로필 수정
-            </button>
-          </div>
-          <BadgesList />
-          <div className="text-lg flex justify-center items-baseline gap-1 mt-8">
-            <p>현재 나무 총</p>
-            <p className="text-xl align-text-bottom font-semibold">{userProfile?.point}</p>
-            <p>그루</p>
+          <div className="text-center mb-4 mx-auto">
+            <img
+              src={userProfile?.profile_img ? `${userProfile?.profile_img}` : 'https://fvkzqozjdtlaogexuuin.supabase.co/storage/v1/object/public/project/userProfileImg/userProfileDefault?t=2023-09-04T01%3A41%3A06.168Z'}
+              alt="profileDefaultImg"
+              width={100}
+              height={100}
+              className="rounded-full inline-block mb-4"
+            />
+            <div className="flex justify-center items-center gap-1 p-2">
+              <p className="font-semibold text-lg">{userProfile?.nickname}</p>
+              <div className="z-100">
+                <RankingGuide />
+              </div>
+            </div>
+            <p className="text-sm opacity-50">{userProfile?.address}</p>
+            <div className="flex justify-center items-center my-3">
+              <Button btnType={'borderBlack'} size={'small'} onClick={handleEditClick}>
+                프로필 수정
+              </Button>
+            </div>
           </div>
         </>
       )}
