+++ src/components/nature-story/index.ts
@@ -1,2 +1,3 @@
 export * from './Content';
+export * from './NaturePage';
 export * from './SideBar';
