+++ src/types/db.type.ts
@@ -30,15 +30,10 @@ export interface User {
   nickname: string | null;
   email: string;
   rank: number;
-  name?: string | null;
 }
 
 export interface AnimalMap {
   [key: string]: string;
-}
-export interface HowToGetTrees {
-  title: string;
-  trees: number;
 }
 
 export interface IdeaComments {
@@ -87,25 +82,4 @@ export interface CertifyPostType {
   img_url: string;
   tags: string;
   challenge_id: string;
-}
-
-export interface Badge {
-  challenge_id: string;
-  completedMission: boolean | null;
-  join_id: string;
-  reviews: number | null;
-  user_id: string;
-  mainChallenge: {
-    badgeUrl: string;
-    challenge_Id: string;
-    content: string;
-    title: string;
-    point: number;
-    limit: number;
-    product: string;
-    startDate: string | null;
-    endDate: string | null;
-    howto: string;
-    animal?: string;
-  } | null;
 }
